app/requirements.txt
# requirements.txt
django==4.2.16
djangorestframework>=3.15.2
python-dotenv>=1.0.1
django-simple-history>=3.7.0
boto3>=1.35.14
watchtower>=3.3.1
djangorestframework-simplejwt>=5.3.1
dj-rest-auth[with_social]>=6.0.0
django-allauth==0.61.1
django-cors-headers>=4.4.0
drf-spectacular[sidecar]>=0.27.2
python-json-logger>=2.0.7
celery[sqs]==5.4.0
uwsgi==2.0.22
psycopg2